Restructure the 'review' folder for clarity.
