+++ src/components/preview/components/Skills.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import {ResumeContext} from "../builder";
+import {ResumeContext} from "../../builder";
 
 const Skills = ({ title, skills }) => {
   const { resumeData, setResumeData } = useContext(ResumeContext);
